# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AjuLabs Shopping API",
    "version": "1.0.0",
    "description": "Marketplace local de Aracaju/SE — consumidores, lojistas e entregadores.\n\nAutenticação via Bearer JWT obtido em /v1/auth/*/login.",
    "contact": {
      "name": "AjuLabs",
      "email": "contato@ajulabs.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local"
    },
    {
      "url": "https://api.ajulabs.com",
      "description": "Produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via /v1/auth/*/login"
      }
    }
  },
  "paths": {
...
    "/v1/avaliacoes": {
      "post": {
        "operationId": "POST_avaliacoes",
        "summary": "Consumidor avalia a loja após pedido entregue",
        "description": "Consumidor avalia a loja após pedido entregue. Atualiza a média de avaliação da loja atomicamente.",
        "tags": [
          "avaliacoes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "201": {
            "description": "Criado com sucesso"
          },
          "400": {
            "description": "INVALID_NOTA: Nota deve ser inteira entre 1 e 5"
          },
          "401": {
            "description": "Token ausente ou inválido"
          },
          "403": {
            "description": "FORBIDDEN: Acesso negado: pedido pertence a outro usuário"
          },
          "404": {
            "description": "NOT_FOUND_LOJA: Loja não encontrada"
          },
          "409": {
            "description": "ALREADY_RATED: Este pedido já foi avaliado"
          },
          "422": {
            "description": "Campos obrigatórios ausentes (spec-validator)"
          },
          "500": {
            "description": "SERVER_ERROR: Erro ao criar avaliação"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "pedidoId",
                  "nota"
                ],
                "properties": {
                  "pedidoId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "uuid, pedido entregue e não avaliado"
                  },
                  "nota": {
                    "type": "number",
                    "description": "int, min 1, max 5"
                  },
                  "comentario": {
                    "type": "string",
                    "description": "max 500 caracteres"
                  }
                }
              },
              "example": {
                "pedidoId": "ped_abc123",
                "nota": 5,
                "comentario": "Entrega super rápida e produto excelente!"
              }
            }
          }
        }
      }
    },
    "/v1/entregador/corridas/{pedidoId}/aceitar": {
      "post": {
        "operationId": "POST_entregador_corridas_pedidoId_aceitar",
        "summary": "Entregador aceita uma oferta de corrida",
        "description": "Entregador aceita uma oferta de corrida. Associa o entregador ao pedido e notifica consumidor e lojista.",
        "tags": [
          "entregador"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "pedidoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "201": {
            "description": "Criado com sucesso"
          },
          "400": {
            "description": "ENTREGADOR_OFFLINE: Entregador precisa estar online"
          },
          "401": {
            "description": "Token ausente ou inválido"
          },
          "500": {
            "description": "SERVER_ERROR: Erro ao aceitar corrida"
          }
        }
      }
    },
    "/v1/entregador/corridas/{pedidoId}/confirmar-entrega": {
      "post": {
        "operationId": "POST_entregador_corridas_pedidoId_confirmar_entrega",
        "summary": "Entregador confirma entrega usando o código de 4 dígitos fornecido pelo consumidor",
        "description": "Entregador confirma entrega usando o código de 4 dígitos fornecido pelo consumidor. Conclui a corrida.",
        "tags": [
          "entregador"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "pedidoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "201": {
            "description": "Criado com sucesso"
          },
          "400": {
            "description": "INVALID_CODE: Código de entrega inválido"
          },
          "401": {
            "description": "Token ausente ou inválido"
          },
          "404": {
            "description": "PEDIDO_NAO_ENCONTRADO: Pedido não encontrado"
          },
          "422": {
            "description": "Campos obrigatórios ausentes (spec-validator)"
          },
          "500": {
            "description": "SERVER_ERROR: Erro ao confirmar entrega"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "codigo"
                ],
                "properties": {
                  "codigo": {
                    "type": "string",
                    "description": "min 1 caractere, 4 dígitos que correspondem ao codigoEntrega do pedido"
                  }
                }
              },
              "example": {
                "pedidoId": "ped_abc123",
                "codigo": "8888"
              }
            }
          }
        }
      }
    },
    "/v1/entregador/heartbeat": {
      "post": {
        "operationId": "POST_entregador_heartbeat",
        "summary": "Heartbeat de localização do entregador no modo \"online idle\" (sem corrida ativa)",
        "description": "Heartbeat de localização do entregador no modo \"online idle\" (sem corrida ativa). Atualiza última posição para oferta de corridas próximas. Chamado a cada ~1 minuto pelo app.",
        "tags": [
          "entregador"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso"
          },
          "201": {
            "description": "Criado com sucesso"
          },
          "400": {
            "description": "INVALID_COORDS: Coordenadas inválidas (lat: -90..90, lng: -180..180)"
          },
          "401": {
            "description": "Token ausente ou inválido"
          },
          "422": {
            "description": "Campos obrigatórios ausentes (spec-validator)"
          },
          "500": {
            "description": "SERVER_ERROR: Erro ao atualizar localização"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "lat",
                  "lng"
                ],
                "properties": {
                  "lat": {
                    "type": "number",
                    "description": "float, min -90, max 90"
                  },
                  "lng": {
                    "type": "number",
                    "description": "float, min -180, max 180"
                  }
                }
              },
              "example": {
                "lat": -10.9776,
                "lng": -37.049
              }
            }
          }
        }
      }
    }
  }
}